Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "__ATMOSCAN_MASTER_V2.0.ino" needs to be inside a sketch folder named "__ATMOSCAN_MASTER_V2.0".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file __ATMOSCAN_MASTER_V2.0.ino to the new folder, leaving behind the other source files. This causes compilation of the sketch to fail:

__ATMOSCAN_MASTER_V2.0.ino:1:67: fatal error: GlobalDefinitions.h: No such file or directory

 #include "GlobalDefinitions.h"

The alternative solution you might be thinking of is to change the sketch name to match the repository name. This would work for people who clone the repository but when GitHub's popular Clone or download > Download ZIP feature is used to download a repository the branch/release/commit name is appended to the folder name, resulting in the sketch and folder names still not matching.